server-py/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
pydantic-settings==2.5.2
sse-starlette==2.1.3
python-dotenv==1.0.1
